searchy/requirements.txt
pillow
tqdm
numpy
scikit-learn
torch
transformers
watchdog
fastapi
uvicorn
# OCR support (macOS Vision framework)
pyobjc-core
pyobjc-framework-Quartz
pyobjc-framework-Vision
# Face recognition
deepface
tf-keras
